docker-debrief/requirements.txt
boto3==1.26.102
exchange_calendars==4.2.5
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
pandas_market_calendars==4.1.4
PyMySQL==1.0.3
python_dateutil==2.8.2
pytz==2023.3
simplejson==3.18.4
mysql-connector-python==8.0.27

#---------------------------------------------
# exchange_calendars==3.5.1
# pandas==1.3.5
# boto3==1.20.15
# simplejson==3.2.0
# numpy==1.21.5
